Upgrade webpack, use webpack-merge to split configuration based on environment
