+++ package.json
@@ -45,15 +45,17 @@
     "gulp-sass": "^4.0.2",
     "mocha": "^6.2.0",
     "style-loader": "^0.16.1",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.8.2"
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.8.1",
+    "webpack-merge": "^4.2.2"
   },
   "scripts": {
     "clean": "rm -rf ./lib",
-    "build": "webpack && gulp sass && cp -R ./src/resources ./lib",
+    "build": "webpack --config webpack.prod.js && gulp sass && cp -R ./src/resources ./lib",
     "lib": "npm run clean && babel src -d lib",
     "prepublish": "npm run lib",
-    "start": "webpack-dev-server",
+    "start": "webpack-dev-server --open --config webpack.dev.js",
     "test": "mocha test/**/*.js --require babel-register --reporter spec || exit 0"
   }
 }
